refactor: rename 'atomEffect' to 'atomComputationEffect'
The new name should highlight the fact the effect runs only on computations
and doesn't trigger computations on its own. If an atom has a computation
effect defined and it's invalidated, it won't be immediately recomputed if
it's not mounted. The effect will... have an effect only after something
triggers recomputation.